requirements.txt
streamlit
groq
python-dotenv
numpy
cryptography
plotly
pandas
requests
flask
flask-cors
